refactor(connect): make functions more extensible and consolidate lp nodes
