+++ src/daft-connect/src/spark_analyzer/literal.rs
@@ -2,6 +2,8 @@ use daft_core::datatypes::IntervalValue;
 use eyre::bail;
 use spark_connect::expression::{literal::LiteralType, Literal};
 
+use crate::not_yet_implemented;
+
 // todo(test): add tests for this esp in Python
 pub fn to_daft_literal(literal: &Literal) -> eyre::Result<daft_dsl::ExprRef> {
     let Some(literal) = &literal.literal_type else {
@@ -9,18 +11,14 @@ pub fn to_daft_literal(literal: &Literal) -> eyre::Result<daft_dsl::ExprRef> {
     };
 
     match literal {
-        LiteralType::Array(_) => bail!("Array literals not yet supported"),
+        LiteralType::Array(_) => not_yet_implemented!("array literals")?,
         LiteralType::Binary(bytes) => Ok(daft_dsl::lit(bytes.as_slice())),
         LiteralType::Boolean(b) => Ok(daft_dsl::lit(*b)),
-        LiteralType::Byte(_) => bail!("Byte literals not yet supported"),
-        LiteralType::CalendarInterval(_) => {
-            bail!("Calendar interval literals not yet supported")
-        }
+        LiteralType::Byte(_) => not_yet_implemented!("Byte literals")?,
+        LiteralType::CalendarInterval(_) => not_yet_implemented!("Calendar interval literals")?,
         LiteralType::Date(d) => Ok(daft_dsl::lit(*d)),
-        LiteralType::DayTimeInterval(_) => {
-            bail!("Day-time interval literals not yet supported")
-        }
-        LiteralType::Decimal(_) => bail!("Decimal literals not yet supported"),
+        LiteralType::DayTimeInterval(_) => not_yet_implemented!("Day-time interval literals")?,
+        LiteralType::Decimal(_) => not_yet_implemented!("Decimal literals")?,
         LiteralType::Double(d) => Ok(daft_dsl::lit(*d)),
         LiteralType::Float(f) => {
             let f = f64::from(*f);
@@ -28,14 +26,14 @@ pub fn to_daft_literal(literal: &Literal) -> eyre::Result<daft_dsl::ExprRef> {
         }
         LiteralType::Integer(i) => Ok(daft_dsl::lit(*i)),
         LiteralType::Long(l) => Ok(daft_dsl::lit(*l)),
-        LiteralType::Map(_) => bail!("Map literals not yet supported"),
+        LiteralType::Map(_) => not_yet_implemented!("Map literals")?,
         LiteralType::Null(_) => {
             // todo(correctness): is it ok to assume type is i32 here?
             Ok(daft_dsl::null_lit())
         }
-        LiteralType::Short(_) => bail!("Short literals not yet supported"),
+        LiteralType::Short(_) => not_yet_implemented!("Short literals")?,
         LiteralType::String(s) => Ok(daft_dsl::lit(s.as_str())),
-        LiteralType::Struct(_) => bail!("Struct literals not yet supported"),
+        LiteralType::Struct(_) => not_yet_implemented!("Struct literals")?,
         LiteralType::Timestamp(ts) => {
             // todo(correctness): is it ok that the type is different logically?
             Ok(daft_dsl::lit(*ts))
