--- src/daft-connect/src/translation/mod.rs
@@ -1,12 +1,0 @@
-//! Translation between Spark Connect and Daft
-
-mod datatype;
-mod expr;
-mod literal;
-mod logical_plan;
-mod schema;
-
-pub use datatype::{to_daft_datatype, to_spark_datatype};
-pub use expr::to_daft_expr;
-pub use literal::to_daft_literal;
-pub use logical_plan::SparkAnalyzer;
